# Optimize Fft