@ebfull 2022-04-21
We added a benchmark in #552 that I ran locally on this PR to assess the impact.

```
Gnuplot not found, using plotters backend
fft/k/3                 time:   [106.66 us 106.87 us 107.08 us]                    
                        change: [+5603.2% +5625.0% +5651.3%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe
fft/k/4                 time:   [300.40 us 301.84 us 303.38 us]                    
                        change: [+8298.5% +8345.8% +8400.4%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
fft/k/5                 time:   [424.64 us 427.30 us 430.03 us]                    
                        change: [+5875.0% +5930.8% +5987.5%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
fft/k/6                 time:   [227.02 us 228.89 us 230.83 us]                    
                        change: [+431.68% +449.11% +463.92%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe
fft/k/7                 time:   [182.95 us 184.51 us 186.17 us]                    
                        change: [+189.29% +200.70% +210.54%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
fft/k/8                 time:   [259.16 us 262.10 us 265.02 us]                    
                        change: [+177.92% +186.59% +195.19%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) low mild
  9 (9.00%) high mild
  1 (1.00%) high severe
fft/k/9                 time:   [365.51 us 369.22 us 373.13 us]                    
                        change: [+128.93% +136.59% +144.39%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
fft/k/10                time:   [561.13 us 568.10 us 575.54 us]                     
                        change: [+95.243% +99.895% +104.69%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
fft/k/11                time:   [848.05 us 852.88 us 857.78 us]                     
                        change: [+62.124% +64.790% +67.511%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe
Benchmarking fft/k/12: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.1s, enable flat sampling, or reduce sample count to 60.
fft/k/12                time:   [1.1982 ms 1.2063 ms 1.2155 ms]                      
                        change: [+31.052% +32.531% +34.144%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) low mild
  7 (7.00%) high mild
  1 (1.00%) high severe
Benchmarking fft/k/13: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.9s, enable flat sampling, or reduce sample count to 50.
fft/k/13                time:   [1.7562 ms 1.7664 ms 1.7769 ms]                      
                        change: [+7.8407% +9.4387% +11.060%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
fft/k/14                time:   [2.6159 ms 2.6315 ms 2.6479 ms]                      
                        change: [-17.393% -16.100% -14.828%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
fft/k/15                time:   [4.0539 ms 4.0786 ms 4.1046 ms]                      
                        change: [-28.847% -27.828% -26.788%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

```

These results suggest two things:

1. at low versions of `k` it is worth having a single thread do the work, as it currently does, due to the overhead of multithreading
2. at high versions of `k` (on my system, `14` or more) it is beneficial to use your patch

There seem to be some tradeoffs in the algorithm that we need to address before we merge this PR.