@NoCtrlZ 2022-04-21
Hi @ebfull 
I found the bottlenecks and fixed.

What I did is separated low degree butterfly arithmetic with condition branch and replaced `par_iter` with `join` because `par_iter` takes overhead as you mentioned.
The following is experiment.
https://github.com/appliedzkp/halo2/pull/55#issue-1195660345

The optimized fft is faster for all k now.
I benched it locally with thread num = 8 and following is the result.

| k | default | recursive |
| ---- | ---- | ---- |
|3| 1.8787 us 1.8841 us 1.8898 us | 510.74 ns 513.45 ns 516.32 ns |
|4| 27.871 us 29.831 us 32.287 us | 13.491 us 13.600 us 13.712 us |
|5| 40.885 us 44.149 us 47.757 us | 17.910 us 17.960 us 18.015 us |
|6| 49.677 us 52.739 us 56.196 us | 29.320 us 29.443 us 29.582 us |
|7| 51.373 us 52.009 us 52.890 us | 41.120 us 41.502 us 41.985 us |
|8| 100.27 us 106.84 us 114.88 us | 54.564 us 54.695 us 54.837 us |
|9| 198.94 us 227.62 us 259.97 us | 83.676 us 84.190 us 84.777 us |
|10| 277.42 us 283.34 us 290.65 us | 144.14 us 145.33 us 146.65 us |
|11| 506.49 us 511.16 us 516.39 us | 269.93 us 272.76 us 275.41 us |
|12| 1.0648 ms 1.0964 ms 1.1335 ms | 534.74 us 539.85 us 544.27 us |
|13| 2.0286 ms 2.0736 ms 2.1223 ms | 1.0254 ms 1.0290 ms 1.0332 ms |
|14| 5.6022 ms 5.8867 ms 6.2248 ms | 2.0972 ms 2.1071 ms 2.1186 ms |
|15| 8.2442 ms 8.4402 ms 8.6559 ms | 4.3206 ms 4.3327 ms 4.3451 ms | 
|16| 19.105 ms 19.842 ms 20.658 ms | 9.1073 ms 9.1606 ms 9.2234 ms |
|17| 37.900 ms 39.631 ms 41.765 ms | 23.751 ms 24.573 ms 25.512 ms |
|18| 73.686 ms 76.518 ms 80.451 ms | 48.620 ms 50.373 ms 52.601 ms |

I would appreciate it if you confirm.
Thank you!