@therealyingtong 2022-04-21
I've locally verified this performance improvement.