@NoCtrlZ 2022-04-22
@daira 
Thank you for the review.
I am going to revert about `log2_floor`.
We also can improve more as I commented but it's a bit complicated and makes code mess.
I have other ideas about optimizations of halo2 which improves performance so much.
We can optimize deeply after major improvement in my opinion.
What do you think?